+++ library/test/test_api_reflection.cpp
@@ -29,14 +29,14 @@ TEST(api_get_profiles_beta, full) {
     uint32_t propertyCount = 0;
     VkResult result0 = vpGetProfiles(&propertyCount, nullptr);
     EXPECT_EQ(VK_SUCCESS, result0);
-    EXPECT_EQ(4, propertyCount);
+    EXPECT_EQ(2, propertyCount);
 
-    propertyCount = 5;
+    propertyCount = 3;
 
     std::vector<VpProfileProperties> properties(propertyCount);
     VkResult result1 = vpGetProfiles(&propertyCount, &properties[0]);
     EXPECT_EQ(VK_SUCCESS, result1);
-    EXPECT_EQ(4, propertyCount);
+    EXPECT_EQ(2, propertyCount);
 
     EXPECT_STREQ(VP_KHR_ROADMAP_2022_NAME, properties[0].profileName);
     EXPECT_EQ(VP_KHR_ROADMAP_2022_SPEC_VERSION, properties[0].specVersion);
@@ -46,7 +46,7 @@ TEST(api_get_profiles_beta, partial) {
     uint32_t propertyCount = 0;
     VkResult result0 = vpGetProfiles(&propertyCount, nullptr);
     EXPECT_EQ(VK_SUCCESS, result0);
-    EXPECT_EQ(4, propertyCount);
+    EXPECT_EQ(2, propertyCount);
 
     propertyCount = 1;
 
@@ -61,7 +61,7 @@ TEST(api_get_profiles_beta, partial) {
 }
 
 TEST(api_get_profile_required_profiles, empty) {
-    const VpProfileProperties profile = {VP_LUNARG_DESKTOP_BASELINE_2022_NAME, 1};
+    const VpProfileProperties profile = {VP_LUNARG_DESKTOP_BASELINE_2023_NAME, 1};
 
     uint32_t count = 0;
     VkResult result0 = vpGetProfileRequiredProfiles(&profile, &count, nullptr);
@@ -95,12 +95,12 @@ TEST(api_get_profile_device_extension_properties, full) {
 }
 
 TEST(api_get_profile_device_extension_properties, partial) {
-    const VpProfileProperties profile = {VP_LUNARG_DESKTOP_BASELINE_2022_NAME, 1};
+    const VpProfileProperties profile = {VP_LUNARG_DESKTOP_BASELINE_2023_NAME, 1};
 
     uint32_t propertyCount = 0;
     VkResult result0 = vpGetProfileDeviceExtensionProperties(&profile, &propertyCount, nullptr);
     EXPECT_EQ(VK_SUCCESS, result0);
-    EXPECT_EQ(37, propertyCount);
+    EXPECT_EQ(41, propertyCount);
 
     propertyCount = 5;
 
@@ -114,7 +114,7 @@ TEST(api_get_profile_device_extension_properties, partial) {
 }
 
 TEST(api_get_profile_instance_extension_properties, full) {
-    const VpProfileProperties profile = {VP_LUNARG_DESKTOP_BASELINE_2022_NAME, 1};
+    const VpProfileProperties profile = {VP_LUNARG_DESKTOP_BASELINE_2023_NAME, 1};
 
     uint32_t propertyCount = 0;
     VkResult result0 = vpGetProfileInstanceExtensionProperties(&profile, &propertyCount, nullptr);
@@ -130,18 +130,17 @@ TEST(api_get_profile_instance_extension_properties, full) {
 }
 
 TEST(api_get_profile_fallbacks, empty) {
-    const VpProfileProperties profile_portability = {VP_LUNARG_DESKTOP_PORTABILITY_2022_NAME, 1};
-    const VpProfileProperties profile_expect = {VP_LUNARG_DESKTOP_BASELINE_2022_NAME, 1};
+    const VpProfileProperties profile = {VP_LUNARG_DESKTOP_BASELINE_2023_NAME, 1};
 
     uint32_t count = 0;
-    VkResult result0 = vpGetProfileFallbacks(&profile_portability, &count, nullptr);
+    VkResult result0 = vpGetProfileFallbacks(&profile, &count, nullptr);
     EXPECT_EQ(VK_SUCCESS, result0);
     EXPECT_EQ(0, count);
 
     count = 1;
 
     std::vector<VpProfileProperties> data(count);
-    VkResult result1 = vpGetProfileFallbacks(&profile_portability, &count, &data[0]);
+    VkResult result1 = vpGetProfileFallbacks(&profile, &count, &data[0]);
     EXPECT_EQ(VK_SUCCESS, result1);
     EXPECT_EQ(0, count);
 }
@@ -176,7 +175,7 @@ TEST(api_get_profile_properties, get_properties2) {
     profileProperties2.sType = VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PROPERTIES_2;
     profileProperties2.pNext = nullptr;
 
-    const VpProfileProperties Profile = {VP_LUNARG_DESKTOP_BASELINE_2022_NAME, 1};
+    const VpProfileProperties Profile = {VP_LUNARG_DESKTOP_BASELINE_2023_NAME, 1};
 
     vpGetProfileProperties(&Profile, &profileProperties2);
 
@@ -189,26 +188,22 @@ TEST(api_get_profile_properties, get_properties2) {
 }
 
 TEST(api_get_profile_structures, get_properties_chain) {
-    VkPhysicalDeviceDescriptorIndexingProperties properties0{};
-    properties0.sType = VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_DESCRIPTOR_INDEXING_PROPERTIES;
-    properties0.pNext = nullptr;
+    VkPhysicalDeviceVulkan12Properties properties1{};
+    properties1.sType = VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VULKAN_1_2_PROPERTIES;
+    properties1.pNext = nullptr;
 
-    VkPhysicalDeviceMultiviewPropertiesKHR properties1{};
-    properties1.sType = VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_MULTIVIEW_PROPERTIES_KHR;
-    properties1.pNext = &properties0;
+    VkPhysicalDeviceVulkan11Properties properties0{};
+    properties0.sType = VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_VULKAN_1_1_PROPERTIES;
+    properties0.pNext = &properties1;
 
-    VkPhysicalDeviceProperties2 properties2{};
-    properties2.sType = VK_STRUCTURE_TYPE_PHYSICAL_DEVICE_PROPERTIES_2;
-    properties2.pNext = &properties1;
+    const VpProfileProperties Profile = {VP_LUNARG_DESKTOP_BASELINE_2023_NAME, 1};
 
-    const VpProfileProperties Profile = {VP_LUNARG_DESKTOP_BASELINE_2022_NAME, 1};
+    vpGetProfileProperties(&Profile, &properties0);
 
-    vpGetProfileProperties(&Profile, &properties2);
-
-    EXPECT_EQ(1048576, properties0.maxUpdateAfterBindDescriptorsInAllPools);
-    EXPECT_EQ(16, properties0.maxPerStageDescriptorUpdateAfterBindSamplers);
-    EXPECT_EQ(134217727, properties1.maxMultiviewInstanceIndex);
-    EXPECT_EQ(6, properties1.maxMultiviewViewCount);
+    EXPECT_EQ(1048576, properties1.maxUpdateAfterBindDescriptorsInAllPools);
+    EXPECT_EQ(16, properties1.maxPerStageDescriptorUpdateAfterBindSamplers);
+    EXPECT_EQ(134217727, properties0.maxMultiviewInstanceIndex);
+    EXPECT_EQ(6, properties0.maxMultiviewViewCount);
 }
 
 TEST(api_get_profile_structures, get_features) {